Make Makefile BSD-Make compatible
Nice issue title 4HEad

Issue was that BSD-Make doesn't support the `$(shell ..)` syntax as this
is a GNU extension. Fix is to define the LIBS variable in a way that it
expands to an inline call to pkg-config instead of expansion in the
variable itself.

Tested on FreeBSD 12.1-RELEASE amd64